CR-67 refactor messages into common
