+++ pom.xml
@@ -138,7 +138,7 @@
     <dependency>
       <groupId>uk.gov.ons.ctp.integration.common</groupId>
       <artifactId>framework</artifactId>
-      <version>0.0.19</version>
+      <version>0.0.21</version>
     </dependency>
 
     <dependency>
