--- src/main/java/uk/gov/ons/ctp/integration/rhsvc/domain/model/CollectionCase.java
@@ -1,20 +1,0 @@
-package uk.gov.ons.ctp.integration.rhsvc.domain.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class CollectionCase {
-
-  private String id;
-  private String caseRef;
-  private String survey;
-  private String collectionExerciseId;
-  private Address address = new Address();
-  private Contact contact = new Contact();
-  private String state;
-  private String actionableFrom;
-}
