--- src/main/java/uk/gov/ons/ctp/integration/rhsvc/domain/model/Contact.java
@@ -1,17 +1,0 @@
-package uk.gov.ons.ctp.integration.rhsvc.domain.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Contact {
-
-  private String title;
-  private String forename;
-  private String surname;
-  private String email;
-  private String telNo;
-}
