--- src/main/java/uk/gov/ons/ctp/integration/rhsvc/domain/model/SurveyLaunchedResponse.java
@@ -1,18 +1,0 @@
-package uk.gov.ons.ctp.integration.rhsvc.domain.model;
-
-import java.util.UUID;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class SurveyLaunchedResponse {
-
-  private String questionnaireId;
-  private UUID caseId;
-  private String agentId;
-}
