+++ src/main/resources/springintegration/uac-event-inbound.xml
@@ -28,9 +28,9 @@
     <int:json-to-object-transformer id="uacEventObject"
                                     input-channel="jsonUacEventMessage"
                                     output-channel="acceptUACEvent"
-                                    type="uk.gov.ons.ctp.integration.rhsvc.message.UACEvent" />
+                                    type="uk.gov.ons.ctp.common.event.model.UACEvent" />
 
-    <int:channel id="acceptUACEvent" datatype="uk.gov.ons.ctp.integration.rhsvc.message.UACEvent" />
+    <int:channel id="acceptUACEvent" datatype="uk.gov.ons.ctp.common.event.model.UACEvent" />
 
     <!-- Start of section to deal with retries and DLQ after max attempts -->
     <int:channel id="uacEventDlqChannel" />
