+++ src/test/java/uk/gov/ons/ctp/integration/rhsvc/integration/SurveyLaunchedIT.java
@@ -1,6 +1,8 @@
 package uk.gov.ons.ctp.integration.rhsvc.integration;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
 import static org.mockito.ArgumentMatchers.any;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static uk.gov.ons.ctp.common.MvcHelper.postJson;
@@ -27,10 +29,10 @@ import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import uk.gov.ons.ctp.common.FixtureHelper;
 import uk.gov.ons.ctp.common.TestHelper;
 import uk.gov.ons.ctp.common.error.RestExceptionHandler;
-import uk.gov.ons.ctp.common.message.Header;
-import uk.gov.ons.ctp.integration.rhsvc.domain.model.SurveyLaunchedResponse;
+import uk.gov.ons.ctp.common.event.model.Header;
+import uk.gov.ons.ctp.common.event.model.SurveyLaunchedEvent;
+import uk.gov.ons.ctp.common.event.model.SurveyLaunchedResponse;
 import uk.gov.ons.ctp.integration.rhsvc.endpoint.RespondentHomeEndpoint;
-import uk.gov.ons.ctp.integration.rhsvc.message.SurveyLaunchedEvent;
 
 /**
  * This is a component test which submits a Post saying that a survey has been launched and uses a
